Reconceptualize refreshScene

+++ src/gameplay.h
@@ -16,7 +16,7 @@
 
 #define BLOCKS 3
 
-s16 scroll[2]; //px
+s32 scroll[2]; //px
 s16 speed[2]; //px per frame
 
 enum GameStates{
